# swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "api 문서",
    "version": "1.0.0",
    "title": "Project_mania",
    "termsOfService": "http://kalin.iptime.org:3100",
    "contact": {
      "name": "YoonJae Song"
    }
  },
  "host": "kalin.iptime.org:3100",
  "basePath": "/",
...
  "schemes": [
    "http"
  ],
  "paths": {
    "/stories": {
      "get": {
        "tags": [
          "story"
        ],
        "summary": "메인화면",
        "description": "",
        "operationId": "addPet",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "description": "여러개의 story가 있는 경우 paging요청으로 추가적으로 story를 요청하기 위해 필요",
            "required": true,
            "type": "integer"
          }
        ],
        "responses": {
          "200": {
            "description": "success",
            "schema": {
              "type": "string"
            }
          }
        }
      },
      "post": {
        "tags": [
          "story"
        ],
        "summary": "Create story",
        "description": "",
        "operationId": "CreateStroy",
        "consumes": [
          "multipart/form-data"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "formData",
            "name": "user_id",
            "required": true,
            "type": "integer"
          },
          {
            "in": "formData",
            "name": "story_title",
            "required": true,
            "type": "string"
          },
          {
            "in": "formData",
            "name": "story_decription",
            "required": true,
            "type": "string"
          },
          {
            "in": "formData",
            "name": "image",
            "type": "file",
            "description": "대표 이미지"
          }
        ],
        "responses": {
          "201": {
            "description": "success"
          }
        }
      }
    },
...
  },
  "securityDefinitions": {
    "petstore_auth": {
      "type": "oauth2",
      "authorizationUrl": "http://petstore.swagger.io/oauth/dialog",
      "flow": "implicit",
      "scopes": {
        "write:pets": "modify pets in your account",
        "read:pets": "read your pets"
      }
    },
    "api_key": {
      "type": "apiKey",
      "name": "api_key",
      "in": "header"
    }
  },
...
}